Updated API to allow filters and extra params
Major refactor of the way http is handled.
Now API's share are single http client instead of create it each time.
The client has hooks to allow the setitng for headers and query parameters
This allows the http logic to be centeralized and allows API Object to have a consistent
way of defining filters.

Auth Objects can now also define Auth Headers as needed.

2Legged auth now supports user context
3Legged auth has expermintal support for implicate auth for JS based apps